Bring back tip.json as subscription.json
